FIX-6 :sparkles: : Controller - ControllerImpl 분리 후, 설명 추가
