+++ src/main/java/com/fixadate/domain/invitation/controller/InvitationController.java
@@ -3,48 +3,26 @@ package com.fixadate.domain.invitation.controller;
 import com.fixadate.domain.invitation.dto.request.InvitationRequest;
 import com.fixadate.domain.invitation.dto.request.InvitationSpecifyRequest;
 import com.fixadate.domain.invitation.dto.response.InvitationResponse;
-import com.fixadate.domain.invitation.service.InvitationService;
+import io.swagger.v3.oas.annotations.Operation;
+import io.swagger.v3.oas.annotations.tags.Tag;
 import jakarta.validation.Valid;
-import lombok.RequiredArgsConstructor;
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
 
-import java.util.List;
+@Tag(name = "InvitationController", description = "InvitationController 입니다.")
+public interface InvitationController {
 
-@RestController
-@RequiredArgsConstructor
-@RequestMapping("/invitation")
-public class InvitationController {
-    private final InvitationService invitationService;
+    @Operation(summary = "초대 등록", description = "초대를 등록합니다.")
+    ResponseEntity<String> registInvitation(@Valid InvitationRequest invitationRequest);
 
-    @PostMapping()
-    public ResponseEntity<String> registInvitation(@Valid @RequestBody InvitationRequest invitationRequest) {
-        String invitationId = invitationService.registInvitation(invitationRequest);
-        return ResponseEntity.status(HttpStatus.CREATED).body(invitationId);
-    }
+    @Operation(summary = "초대 조회", description = "초대 코드를 입력하여 초대에 응답합니다.")
+    ResponseEntity<?> checkInvitationById(String id);
 
-    @GetMapping() // 초대 코드를 입력해서 초대에 응답하는 로직
-    public ResponseEntity<?> checkInvitationById(@RequestParam String id) {
-        return ResponseEntity.ok(invitationService.getInvitationById(id));
-    }
+    @Operation(summary = "초대 존재 여부 확인", description = "중복 방지를 위해 이미 생성된 초대장이 있는지 조회합니다.")
+    ResponseEntity<InvitationResponse> isInvitationExist(Long datesId);
 
-    //중복 방지를 위해 이미 생성된 초대장이 있는지 조회하는 로직
-    @GetMapping("/check")
-    public ResponseEntity<InvitationResponse> isInvitationExist(@RequestParam Long datesId) {
-        return ResponseEntity.ok(invitationService.getInvitationFromTeamId(datesId));
-    }
+    @Operation(summary = "멤버 초대", description = "특정 멤버를 팀에 초대합니다.")
+    ResponseEntity<?> inviteMemberToTeams(@Valid InvitationSpecifyRequest requestDto);
 
-    @PostMapping("/specify")
-    public ResponseEntity<?> inviteMemberToTeams(@Valid @RequestBody InvitationSpecifyRequest requestDto) {
-        invitationService.inviteSpecifyMember(requestDto);
-        return ResponseEntity.status(HttpStatus.CREATED).build();
-    }
-
-    @GetMapping("/specify")
-    public ResponseEntity<?> getSpecifyInvitationByTeamId(@RequestParam Long teamId) {
-        List<InvitationResponse> responses = invitationService.getInvitationResponseFromTeamId(teamId);
-        return ResponseEntity.ok(responses);
-    }
+    @Operation(summary = "팀별 초대 조회", description = "특정 팀에 대한 초대를 조회합니다.")
+    ResponseEntity<?> getSpecifyInvitationByTeamId(Long teamId);
 }
